:recycle: NewDB()を移動
